Fix: new windows problem fixed

+++ StreamAndConnexion/ClientV2/Client.cpp
@@ -234,7 +234,7 @@ int Client::ReceiveData() {
         if (data.msg_id > last_rcv_id) {
           int diff = static_cast<int>(data.msg_id - last_rcv_id);
           int size = static_cast<int>(received_packages.size());
-          received_packages <<= std::min(diff, size);
+          received_packages <<= static_cast<size_t>(std::min(diff, size));
           received_packages[0] = true;
           last_rcv_id = data.msg_id;
         }else {
